+++ packages/google_sign_in/google_sign_in_web/lib/google_sign_in_web.dart
@@ -4,14 +4,31 @@
 
 import 'dart:async';
 import 'dart:html' as html;
+import 'dart:ui' as ui;
 
 import 'package:flutter/foundation.dart' show visibleForTesting, kDebugMode;
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart' show PlatformException;
+import 'package:flutter/widgets.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:google_identity_services_web/loader.dart' as loader;
 import 'package:google_sign_in_platform_interface/google_sign_in_platform_interface.dart';
 
+import 'src/button_configuration.dart' show GSIButtonConfiguration;
+import 'src/dom.dart';
+import 'src/flexible_size_html_element_view.dart';
 import 'src/gis_client.dart';
+
+// Export the configuration types for the renderButton method.
+export 'src/button_configuration.dart'
+    show
+        GSIButtonConfiguration,
+        GSIButtonLogoAlignment,
+        GSIButtonShape,
+        GSIButtonSize,
+        GSIButtonText,
+        GSIButtonTheme,
+        GSIButtonType;
 
 /// The `name` of the meta-tag to define a ClientID in HTML.
 const String clientIdMetaName = 'google-signin-client_id';
@@ -33,6 +50,8 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
         .querySelector(clientIdMetaSelector)
         ?.getAttribute(clientIdAttributeName);
 
+    _registerButtonFactory();
+
     if (debugOverrideLoader) {
       _jsSdkLoadedFuture = Future<bool>.value(true);
     } else {
@@ -40,7 +59,11 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
     }
   }
 
+  // A future that completes when the JS loader is done.
   late Future<void> _jsSdkLoadedFuture;
+  // A future that completes when the `init` call is done.
+  final Completer<void> _initDone = Completer<void>();
+  // A (synchronous) marker to assert that `init` has been called.
   bool _isInitCalled = false;
 
   // The instance of [GisSdkClient] backing the plugin.
@@ -62,7 +85,8 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
   @visibleForTesting
   Future<void> get initialized {
     _assertIsInitCalled();
-    return _jsSdkLoadedFuture;
+    return Future.wait<void>(
+        <Future<void>>[_jsSdkLoadedFuture, _initDone.future]);
   }
 
   /// Stores the client ID if it was set in a meta-tag of the page.
@@ -110,6 +134,8 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
         'Check https://developers.google.com/identity/protocols/googlescopes '
         'for a list of valid OAuth 2.0 scopes.');
 
+    _isInitCalled = true; // Mark `init` ASAP before going async.
+
     await _jsSdkLoadedFuture;
 
     _gisClient = overrideClient ??
@@ -120,19 +146,54 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
           loggingEnabled: kDebugMode,
         );
 
-    _isInitCalled = true;
+    _initDone.complete(); // Signal that `init` is fully done.
+  }
+
+  // Register a factory for the Button HtmlElementView.
+  void _registerButtonFactory() {
+    // ignore: avoid_dynamic_calls, undefined_prefixed_name
+    ui.platformViewRegistry.registerViewFactory(
+      'gsi_login_button',
+      (int viewId) {
+        final DomElement element = createDomElement('div');
+        element.setAttribute('style',
+            'width: 100%; height: 100%; overflow: hidden; display: flex; flex-wrap: wrap; align-content: center; justify-content: center;');
+        element.id = 'sign_in_button_$viewId';
+        return element;
+      },
+    );
+  }
+
+  /// Render the GSI button web experience.
+  Widget renderButton({GSIButtonConfiguration? configuration}) {
+    final GSIButtonConfiguration config =
+        configuration ?? GSIButtonConfiguration();
+    return FutureBuilder<void>(
+      key: Key(config.hashCode.toString()),
+      future: initialized,
+      builder: (BuildContext context, AsyncSnapshot<void> snapshot) {
+        if (snapshot.hasData) {
+          return FlexHtmlElementView(
+              viewType: 'gsi_login_button',
+              onPlatformViewCreated: (int viewId) {
+                final DomElement? element =
+                    domDocument.querySelector('#sign_in_button_$viewId');
+                assert(element != null,
+                    'Cannot render GSI button. DOM is not ready!');
+                _gisClient.renderButton(element!, config);
+              });
+        }
+        return const Text('Getting ready');
+      },
+    );
   }
 
   @override
   Future<GoogleSignInUserData?> signInSilently() async {
     await initialized;
 
-    // Since the new GIS SDK does *not* perform authorization at the same time as
-    // authentication (and every one of our users expects that), we need to tell
-    // the plugin that this failed regardless of the actual result.
-    //
-    // However, if this succeeds, we'll save a People API request later.
-    return _gisClient.signInSilently().then((_) => null);
+    // The new user is being injected from the `userDataEvents` Stream.
+    return _gisClient.signInSilently(); //.then((_) => null);
   }
 
   @override
@@ -201,4 +262,15 @@ class GoogleSignInPlugin extends GoogleSignInPlatform {
 
     return _gisClient.requestScopes(scopes);
   }
+
+  @override
+  Future<bool> canAccessScopes(String? accessToken, List<String> scopes) async {
+    await initialized;
+
+    return _gisClient.canAccessScopes(accessToken, scopes);
+  }
+
+  @override
+  Stream<GoogleSignInUserData?>? get userDataEvents =>
+      _gisClient.userDataEvents;
 }