+++ packages/google_sign_in/google_sign_in_web/lib/src/gis_client.dart
@@ -11,6 +11,9 @@ import 'package:google_sign_in_platform_interface/google_sign_in_platform_interf
 import 'package:js/js.dart';
 import 'package:js/js_util.dart';
 
+import 'button_configuration.dart'
+    show GSIButtonConfiguration, convertButtonConfiguration;
+import 'dom.dart';
 import 'people.dart' as people;
 import 'utils.dart' as utils;
 
@@ -24,8 +27,9 @@ class GisSdkClient {
     required String clientId,
     bool loggingEnabled = false,
     String? hostedDomain,
-  }) : _initialScopes = initialScopes {
-    if (loggingEnabled) {
+  })  : _initialScopes = initialScopes,
+        _loggingEnabled = loggingEnabled {
+    if (_loggingEnabled) {
       id.setLogLevel('debug');
     }
     // Configure the Stream objects that are going to be used by the clients.
@@ -45,6 +49,12 @@ class GisSdkClient {
     );
   }
 
+  void _logIfEnabled(String message, [List<Object?>? more]) {
+    if (_loggingEnabled) {
+      domConsole.info('[google_sign_in_web] $message', more);
+    }
+  }
+
   // Configure the credential (authentication) and token (authorization) response streams.
   void _configureStreams() {
     _tokenResponses = StreamController<TokenResponse>.broadcast();
@@ -52,13 +62,25 @@ class GisSdkClient {
     _tokenResponses.stream.listen((TokenResponse response) {
       _lastTokenResponse = response;
     }, onError: (Object error) {
+      _logIfEnabled('Error on TokenResponse:', <Object>[error.toString()]);
       _lastTokenResponse = null;
     });
     _credentialResponses.stream.listen((CredentialResponse response) {
       _lastCredentialResponse = response;
     }, onError: (Object error) {
+      _logIfEnabled('Error on CredentialResponse:', <Object>[error.toString()]);
       _lastCredentialResponse = null;
     });
+
+    // In the future, the userDataEvents could propagate null userDataEvents too.
+    userDataEvents = _credentialResponses.stream
+        .map(utils.gisResponsesToUserData)
+        .handleError(
+      (Object error) {
+        _logIfEnabled('Removing error from `userDataEvents`:',
+            <Object>[error.toString()]);
+      },
+    );
   }
 
   // Initializes the `id` SDK for the silent-sign in (authentication) client.
@@ -185,6 +207,14 @@ class GisSdkClient {
     }
   }
 
+  /// Calls `id.renderButton` on [parent] with the given [options].
+  Future<void> renderButton(
+    Object parent,
+    GSIButtonConfiguration options,
+  ) async {
+    return id.renderButton(parent, convertButtonConfiguration(options)!);
+  }
+
   /// Starts an oauth2 "implicit" flow to authorize requests.
   ///
   /// The new GIS SDK does not return user authentication from this flow, so:
@@ -273,7 +303,14 @@ class GisSdkClient {
   ///
   /// Keeps the previously granted scopes.
   Future<bool> requestScopes(List<String> scopes) async {
+    // If we already know the user, use their `email` as a `hint`, so they don't
+    // have to pick their user again in the Authorization popup.
+    final GoogleSignInUserData? knownUser =
+        utils.gisResponsesToUserData(_lastCredentialResponse);
+
     _tokenClient.requestAccessToken(OverridableTokenClientConfig(
+      prompt: knownUser == null ? 'select_account' : '',
+      hint: knownUser?.email,
       scope: scopes.join(' '),
       include_granted_scopes: true,
     ));
@@ -282,6 +319,22 @@ class GisSdkClient {
 
     return oauth2.hasGrantedAllScopes(_lastTokenResponse!, scopes);
   }
+
+  /// Checks if the passed-in `accessToken` can access all `scopes`.
+  ///
+  /// This validates that the `accessToken` is the same as the last seen
+  /// token response, and uses that response to check if permissions are
+  /// still granted.
+  Future<bool> canAccessScopes(String? accessToken, List<String> scopes) async {
+    if (accessToken != null && _lastTokenResponse != null) {
+      if (accessToken == _lastTokenResponse!.access_token) {
+        return oauth2.hasGrantedAllScopes(_lastTokenResponse!, scopes);
+      }
+    }
+    return false;
+  }
+
+  final bool _loggingEnabled;
 
   // The scopes initially requested by the developer.
   //
@@ -300,6 +353,9 @@ class GisSdkClient {
   // The last-seen credential and token responses
   CredentialResponse? _lastCredentialResponse;
   TokenResponse? _lastTokenResponse;
+
+  /// The Stream of asynchronous User Authentication events.
+  late Stream<GoogleSignInUserData?> userDataEvents;
 
   // If the user *authenticates* (signs in) through oauth2, the SDK doesn't return
   // identity information anymore, so we synthesize it by calling the PeopleAPI